+++ empire/server/modules/csharp/Inject_BOF.Covenant.py
@@ -48,7 +48,7 @@ class Module:
 
         script_file = (
             main_menu.installPath
-            + "/csharp/Covenant/Data/Tasks/CSharp/Compiled/"
+            + "/Empire-Compiler/EmpireCompiler/Data/Tasks/CSharp/Compiled/"
             + (params["DotNetVersion"]).lower()
             + "/"
             + file_name