+++ empire/server/plugins/basic_reporting/basic_reporting.py
@@ -29,7 +29,7 @@ class Plugin(BasePlugin):
         db = kwargs["db"]
         input = f'Generating reports for: {command["report"]}'
         plugin_task = models.PluginTask(
-            plugin_id=self.info["Name"],
+            plugin_id=self.info.name,
             input=input,
             input_full=input,
             user_id=user.id,