+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Plugin execute function must take `**kwargs`
 -   Plugin name is now based on the name in the `plugin.yaml` file instead of the filename
 -   Plugins now contain `main_menu` instead of `mainMenu`
+-   Moved C# compiler to a submodule.
+-   All C# module code has been moved as submodules of Empire-Compiler
 
 ### Deprecated